+++ CHANGELOG.md
@@ -5,6 +5,9 @@
     - Just provide `entities` and `identifiers` options to connection spec to
     convert column names.
 - Remove deprecated functions in `metosin.dates`
+- Add `metosin.jdbc.java-time` to provide time conversions for Clojure.java.jdbc
+- Add `metosin.transit.java-time` to provide readers and writers for Transit
+using Java time
 
 ## 0.5.0 (2018-08-31)